requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
requests

# LangChain Ecosystem
langchain
langchain-community
langchain-text-splitters
langchain-ollama
langchain-google-genai

# Vector Database & Document Processing
faiss-cpu
pypdf

# Google SDK
google-genai